feat: add recipe add-function
## Description 

I'm not sure that this is the best way to go about it - and as we, @lwjohnst86 , talked about we might want to make a tool for tasks like this at some point. But maybe this is fine for now? 

At least I find this helpful - instead of having to create the function file and corresponding test file by hand.