feat: 🧑‍💻 add recipe `add-function` (#2)
## Description 

I'm not sure that this is the best way to go about it - and as we,
@lwjohnst86 , talked about we might want to make a tool for tasks like
this at some point. But maybe this is fine for now?

At least I find this helpful - instead of having to create the function
file and corresponding test file by hand.

With this, the command `just add-function sprout core hello` creates
`sprout/core/hello.py` and the corresponding test file
`tests/core/test_hello.py`

## Reviewer Focus

<!-- Select quick/in-depth as necessary -->
This PR needs a quick review.