fix: :hammer: update test snippet to given, when, then (#29)
Since we have moved away from AAA.
Alternatively, this can be removed. I'm not sure anyone uses it. 

(Saw it this morning during the status meeting, when looking at the
`examples` repo and was reminded that I made this change a while ago
without pushing it)

<!-- Select quick/in-depth as necessary -->
This PR needs a quick review.