docs: :page_facing_up: match author style of LICENSE to the template (#164)
# Description

Using this style since a template generator (to my understanding) is
less software and more of an opinion on a structured set of
files/folders, so doesn't exactly fall under the same copyright rules.

Plus updated links to the license to refer to it at the root folder.

This PR needs a quick review.